nix users module: write an actual callUserCfgFn functor
The NixOS module system was smart enough to figure a functor of
singletons, but when a merge is required it fails to complete it.

I'm still piggy-backing the functor in home-manager and there's no type
checking. Since nix does not have cofunctors, I'd have to assert during
the merge. I think this is useless and I'll cross that bridge when I'm
proven wrong.

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>