nix modules util: add function renderDotfile
From it's docstring:

Replaces every occurrence of @varName@ in a copy of every file in path,
where varName is any variable in the specified attr set. If path is a file
only that file will be copied and processed.
The variable @destDir@ will be automatically added with value equal to the
derivation path.
Variables may start with lowercase and contain only alphabetic characters.

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>