mailcap: Add entry for text/plain
In NixOS 20.09, alot was updated to version 0.9.1.

Since alot 0.9.1 it is possible to specify a mailcap for text/plain.
This was delegating it to xdg-open, which in turns delegates it to
nvim. which probably tried to take control of the tty and crashed alot
(see https://github.com/pazz/alot/issues/1544). This adds an explicit
entry for text/plain.

Mind that the order in which alot matches this file seems to be important.
This is probably a bug, but fixed here as a workaround.

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>