Minerva: Add trusted user to allow using flakes for deployment
Remote building on flakes does not seem to be working.
Copying the derivation is restricted by default since usually
the build host does not sign the built derivations (or the signing
key is not trusted by the receiving host).

Adding the user used to copy the closure to the remote machine allows
copying untrusted paths to the remote machine.

The process of deploying to a remote machine is still somewhat painful:

```
$ nix build .#nixosConfigurations.Minerva.config.system.build.toplevel -L
$ nix copy --to ssh://minerva.int ./result
$ ssh minerva.int -- sudo $(readlink -f result)/bin/switch-to-configuration switch
```

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>