Heimdaalr: Increase connections per nginx worker
From the nginx documentation:

> It should be kept in mind that this number includes all connections (e.g.
> connections with proxied servers, among others), not only connections with
> clients. [...]

The default value is 512, moving it to 4096 gives 2k+ connections per
side (clients, upstreams).

Moreover,

> [...] Another consideration is that the actual number of simultaneous
> connections cannot exceed the current limit on the maximum number of open
> files, which can be changed by worker_rlimit_nofile.

The current hard limit is high off the sky:

```console
 Heimdaalr:~ sudo -u nginx bash -c 'ulimit -Hn'
524288
```

The soft limit however, is less than this number.
Users are allowed to not only go over the soft limit, but also modify it
up to the hard limit. I'm curious to see what nginx does.

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>