Add 'firewall' module featuring dynamic security zones
This module implements the concept of 'networks' by associating
interfaces to a set of rules.

The interface ipv6 adresses are monitored and machines in the same
neighborhood (cidr) are applied the specified rules.

The monitoring works by subscribing to RTNETLINK for inet6 address
changes (via ip monitor) and trigerring a change event in udev.
This triggers a shell script that takes the ipv6 addresses of the
interface and creates iptables rules matching any computer withing the
same segment.

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>