ups/roosemberth: Enable zsh system-wide
My zsh configuration is managed through home-manager but aparently
nixpkgs now does a bunch of stuff to provide a usable shell when the
configuration is lacking; so much that users without this shenanigans
wouldn't be able to login. This is inconsequential to my user's shell,
but may be nice when I login as a different user.

Signed-off-by: Roosembert Palacios <roosemberth@posteo.ch>