Move NixOS-hardware settings along with module loader
These settings are machine and hardware-specific. The usual abstraction
(include the module when available) used by other modules and module-specific
options cannot be used here, because modules in NixOS-hardware do not
have the usual behaviour to do nothing unless an option is enabled.
Instead, they must be carefully included on a per-machine and hardware basis.